fix some bug

+++ app/Http/Controllers/Controller.php
@@ -25,10 +25,16 @@ class Controller extends BaseController
     }
     protected function updateImage($model){
         if($model->image){
-            $this->deleteFile($model->image->url??null);
-            $model->image()->update(['url'=>request()->file('image')->store('images','public')]);
+            if(request()->hasFile('image')){
+                $this->deleteFile($model->image->url??null);
+                $model->image()->update(['url'=>request()->file('image')->store('images','public')]);
+            }
+           
+
         }else{
-            $model->image()->create(['url'=>request()->file('image')->store('images','public')]); 
+            if(request()->hasFile('image')){
+            $model->image()->create(['url'=>request()->file('image')->store('images','public')]);
+            } 
         }
     }
     protected function storeImage($model){
